+++ RELEASE_NOTES.md
@@ -1,3 +1,23 @@
+### 0.8.1 (2023-12-30)
+
+- BREAKING CHANGE: `DemoParser.Start` is now `DemoParser.ReadAllAsync`
+- Added ability to read tick-by-tick with `DemoParser.StartReadingAsync` and `DemoParser.MoveNextAsync`
+- Added comparison operators to `DemoTick` and `GameTick`
+
+### 0.7.1 (2023-12-23)
+
+- Read CDemoFileInfo in seekable streams (available on complete demo recordings only)
+
+### 0.6.1 (2023-12-23)
+
+- Only read the first entity snapshot in GOTV demos (~10% perf improvement)
+- Added `DemoFile.IsGotv`, indicating whether the demo was recorded by GOTV
+- Fix pawn position jumping to multiples of 1024 for a single tick (https://github.com/saul/demofile-net/issues/27)
+
+### 0.5.1 (2023-12-19)
+
+- Fix decode exception caused by .NET runtime bug (https://github.com/dotnet/runtime/issues/96174)
+
 ### 0.4.1 (2023-12-09)
 
 - Update to schema v13980